Cleaned up available autos and renamed to follow convention.
