+++ src/main/java/frc/robot/Autonomous/Mid_B_2_3.java
@@ -1,136 +1,149 @@
 package frc.robot.Autonomous;
+
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
-import frc.robot.Enums.IntakeState;
-import frc.robot.Enums.JukeboxEnum;
-import frc.robot.Subsystems.Drivetrain;
+import frc.robot.Enums.*;
+import frc.robot.Subsystems.*;
+import frc.robot.Utilities.PathFollower;
 import frc.robot.Subsystems.Intake;
 import frc.robot.Subsystems.Jukebox;
-import frc.robot.Utilities.PathFollower;
 
-public class bottomAutoMode extends AutonomousMode{
-
-    private PathFollower _pathFollower;
+public class Mid_B_2_3 extends AutonomousMode{
+    private PathFollower _pathFollower;    
+    private Drivetrain _drivetrain;
     private Intake _intake;
     private Jukebox _jukebox;
-    private Drivetrain _drivetrain;
     private int _count;
-    private int _loopCounter;
 
-    public bottomAutoMode()
-    {
+    public Mid_B_2_3(){
         _drivetrain = Drivetrain.getInstance();
         _intake = Intake.getInstance();
         _jukebox = Jukebox.getInstance();
-        _pathFollower = new PathFollower("Bottom Path1");
-        resetLoopTimer();
+        _pathFollower = new PathFollower("MID B 2 3");
+        _intake = Intake.getInstance();
+        _jukebox = Jukebox.getInstance();
     }
 
     @Override
-    public void execute()
-    {
-        
-        switch(_count)
-        {
+    public void execute(){
+        switch (_count) {
+            // preps first shot
             case 0:
                 _jukebox.setState(JukeboxEnum.PREP_SPEAKER);
                 nextStep();
                 break;
+                // scores first shot
             case 1:
-                if (_jukebox.isReadyToScore()){
-                    _jukebox.setState(JukeboxEnum.SCORE);
-                    nextStep();
-                }
-                break;
-            case 2:
-                if (!_jukebox.hasNote())
-                {
-                    _jukebox.setState(JukeboxEnum.IDLE);
-                    // Start Path to Note 5 mid
-                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotation());
-                    nextStep();
-
-                }
-                break;
-            case 3:
-                // Follow Path to Note 5 mid
-                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
-                if (_pathFollower.isPathFinished())
-                {
-                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset());
-                    _drivetrain.drive(new ChassisSpeeds());
-                    nextStep();
-                }
-                break;
-            case 4:
-                // Follow Path back to note 5
-                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
-                if (_pathFollower.isPathFinished())
-                {
-                    _drivetrain.drive(new ChassisSpeeds());
-                    nextStep();
-                }
-                break;
-            case 5:
-                if (_jukebox.hasNote()) {
-                    _jukebox.setState(JukeboxEnum.PREP_SPEAKER);                
-                    nextStep();
-                }
-                break;
-            case 6:
                 if (_jukebox.isReadyToScore()) {
                     _jukebox.setState(JukeboxEnum.SCORE);
                     nextStep();
                 }
                 break;
-            case 7:
+                // resets robot
+            case 2:
                 if (!_jukebox.hasNote()) {
                     _jukebox.setState(JukeboxEnum.IDLE);
-                    // start bottom to note 4 center
                     _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset());
                     nextStep();
                 }
                 break;
-            case 8:
-                // Follow Path to Note 4
+            // intakes b
+            case 3:
                 _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
-                if (_pathFollower.isPathFinished())
-                {
-                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset());
+                if (_pathFollower.isPathFinished()){
                     _drivetrain.drive(new ChassisSpeeds());
                     nextStep();
                 }
                 break;
-            case 9:
-                // Follow Path back to Goal
-                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
-                if (_pathFollower.isPathFinished())
-                {
-                    _drivetrain.drive(new ChassisSpeeds());
-                    nextStep();
-                }
-                break;
-            case 10:
+            // preps at b
+            case 4:
                 if (_jukebox.hasNote()) {
                     _jukebox.setState(JukeboxEnum.PREP_SPEAKER);                
                     nextStep();
                 }
-            case 11:
+                break;
+            // shoots
+            case 5:
                 if (_jukebox.isReadyToScore()) {
                     _jukebox.setState(JukeboxEnum.SCORE);
                     nextStep();
                 }
                 break;
-            case 12:
-                if (!_jukebox.hasNote()){
+            // resets and starts path to 3
+            case 6:
+                if (!_jukebox.hasNote()) {
                     _jukebox.setState(JukeboxEnum.IDLE);
+                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset());
+                    nextStep();
+                }
+                break;
+            // drives to 2
+            case 7:
+                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
+                if (_pathFollower.isPathFinished()){
+                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset()); 
                     _drivetrain.drive(new ChassisSpeeds());
                     nextStep();
                 }
-            default:
-                _drivetrain.drive(new ChassisSpeeds());
+                break;
+            // drives back to B
+            case 8:
+                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
+                if (_jukebox.hasNote()) {
+                    _jukebox.setState(JukeboxEnum.PREP_SPEAKER);
+                }
+                if (_pathFollower.isPathFinished()){
+                    _drivetrain.drive(new ChassisSpeeds());
+                    nextStep();
+                }
+                break;
+            // scores at b
+            case 9:
+                if (_jukebox.isReadyToScore()) {
+                    _jukebox.setState(JukeboxEnum.SCORE);
+                    nextStep();
+                }
+                break;
+            // resets after shot
+            case 10:
+                if (!_jukebox.hasNote()) {
+                    _jukebox.setState(JukeboxEnum.IDLE);
+                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset());
+                    nextStep();
+                }
+                break;
+            // drives to 3
+            case 11:
+                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
+                if (_pathFollower.isPathFinished()){
+                    _pathFollower.startNextPath(new ChassisSpeeds(), _drivetrain.getGyroRotationWithOffset()); 
+                    _drivetrain.drive(new ChassisSpeeds());
+                    nextStep();
+                }
+                break;
+            // drives back to B
+            case 12:
+                _drivetrain.drive(_pathFollower.getPathTarget(_drivetrain.getPose()));
+                if (_jukebox.hasNote()) {
+                    _jukebox.setState(JukeboxEnum.PREP_SPEAKER);
+                }
+                if (_pathFollower.isPathFinished()){
+                    _drivetrain.drive(new ChassisSpeeds());
+                    nextStep();
+                }
+                break;
+            // shoots at b
+            case 13:
+                if (_jukebox.isReadyToScore()) {
+                    _jukebox.setState(JukeboxEnum.SCORE);
+                    nextStep();
+                }
+                break;
+            case 14:
+                if (!_jukebox.hasNote()) {
+                    _jukebox.setState(JukeboxEnum.IDLE);
+                }
                 break;
         }
-        _loopCounter++;
     }
 
     @Override
@@ -138,28 +151,16 @@ public class bottomAutoMode extends AutonomousMode{
 
     @Override
     public boolean isComplete(){
-        return _count >= 13;
+        return _count >= 15;
     }
     @Override
     public void initialize(){
         var startingPose = _pathFollower.getStartingPose();
+        _intake.setWantedState(IntakeState.INTAKE);
         _drivetrain.setStartingPose(startingPose);
         _count = 0;
-        _intake.setWantedState(IntakeState.INTAKE);
-        _jukebox.setState(JukeboxEnum.IDLE);
-        _loopCounter = 0;
     }
     private void nextStep() {
         _count++;
-    }
-    private void resetLoopTimer()
-    {
-        _loopCounter = 0;
-    }
-    private boolean hasElapsed(double seconds)
-    {
-        // There are 50 loops per second (0.02 seconds)
-        // There are 750 loops in 15 seconds
-        return ((double)_loopCounter / 50) >= seconds;
     }
 }
